refactor: move seed script data to scripts/seed-data/
Move files only used by seed script out of src/config:
- nouns-home.ts, clanker-home.ts (navPage layouts)
- exploreTabs/*.json (preloaded directory data)
- exploreTabDirectoryData.ts (parser utility)
- nouns-assets/ (brand assets for ImgBB upload)

Move clanker logo to public/images/ (used by FidgetPickerModal)

Delete unused clanker/ config directory entirely

src/config/ now only contains:
- Runtime config loading (loaders/)
- Type definitions (systemConfig.ts)
- Theme (nouns.theme.ts, shared/themes.ts)
- Initial space creators (nouns/initialSpaces/)
