+++ src/common/components/organisms/FidgetPickerModal.tsx
@@ -12,7 +12,6 @@ import {
 } from "@/common/types/fidgetOptions";
 import { Input } from "../atoms/input";
 import Image from "next/image";
-import clankerLogo from "@/config/clanker/assets/nobg.png";
 
 import { Search } from "lucide-react";
 
@@ -60,7 +59,7 @@ const TAG_CONFIG: Record<string, { color: string; icon: string | any; displayNam
   'tiktok': { color: 'bg-black text-white', icon: '🎵' },
   'skateboarding': { color: 'bg-orange-100 text-orange-800', icon: '🛹' },
   'aerodrome': { color: 'bg-blue-100 text-blue-800', icon: '✈️' },
-  'clanker': { color: 'bg-purple-100 text-purple-800', icon: clankerLogo, displayName: 'Clanker' },
+  'clanker': { color: 'bg-purple-100 text-purple-800', icon: '/images/clanker-logo.png', displayName: 'Clanker' },
   'scheduling': { color: 'bg-blue-100 text-blue-800', icon: '📅' },
   'presentations': { color: 'bg-gray-100 text-gray-800', icon: '📊' },
   'networking': { color: 'bg-blue-100 text-blue-800', icon: '👥' },
