--- src/config/clanker/index.ts
@@ -1,2 +1,0 @@
-// Export home page config (used by seed script)
-export { clankerHomePage } from './clanker.home';
